==> analysis/02_clean_source.ipynb <==
# Notebook 2: Clean Source to Extract Transcripts

## Introduction

Now that I've extracted the source for each page of transcripts, I need to extract the actual transcript text. There is unfortunately no overarching, easily digestible method to get a the cleaned transcript, and I relied mostly on trial and error to get the text I wanted. I'm certain there are places where the transcripts are not 100% clean, and I either tried to address this at some later point with more data cleaning, or let it go, because the analysis makes sense with the data in its current state.

import pandas as pd
import numpy as np
import time, os
import re
import pickle
import string

## Clean deBlasio Transcripts

### Extracting transcripts from source code

This process is messy, and I ended up using two different cleaning functions - a second one to catch the ones missed by the first function. 

with open('../data/bdbsource_519.pickle', 'rb') as read_file:
    bdbsource_519 = pickle.load(read_file)
with open('../data/bdblinks_519.pickle', 'rb') as read_file:
    bdblinks_519 = pickle.load(read_file) 

def get_text(source_object):
    text_list = []
    for s in source_object:
        text = s.find_all('p')
        text_list.append(text)
    return text_list

transcript_length = []
for i in get_text(bdbsource_519):
    transcript_length.append(len(i))
transcript_length_array = np.array(transcript_length)

first_transcripts = get_text(bdbsource_519)

# These are the indices that need to be re-extracted with get_text2
indices = list(np.argwhere(transcript_length_array == 1).flatten())

def get_text2(source_object):
    text_list = []
    for s in source_object:
        text = s.find('p').parent
        text_list.append(text)
    return text_list

# Extract a better transcript from the bad indices
better_transcripts = []
for i in indices:
    better_transcripts.append(get_text2(bdbsource_519[i]))

#replace first transcripts with better_transcripts
for (indices, better_transcripts) in zip(indices, better_transcripts):
    first_transcripts[indices] = better_transcripts

### Additional Cleaning Steps

I also wanted to remove html tags, extract the date of the speech, and, in instances where the speech was actually an interview, take only the parts where Mayor de Blasio was speaking.

def remove_html_tags(text):
    clean = re.compile('<.*?>')
    return re.sub(clean, '', text)

def clean_bdb(transcript_list, link_list):
    date = []
    text = []
    for i in transcript_list:
        cleaned = remove_html_tags(str(i))
        if re.search('\[\\n(.+)\\n(.+)\]', cleaned) is not None and re.search('\[\\n(.+)\\n(.+)\]', cleaned) is not None:
            date_clean = re.search('\[\\n(.+)\\n(.+)\]', cleaned).group(1)
            date.append(date_clean)
            text_clean = re.search('\[\\n(.+)\\n(.+)\]', cleaned).group(2)
            text.append(text_clean)
        else:
            date_clean = re.search('\\n\\n(.+20.{2})\\n(.+)\\n\\ufeff', cleaned).group(1)
            date.append(date_clean)
            text_clean = re.search('\\n\\n(.+20.{2})\\n(.+)\\n\\ufeff', cleaned).group(2)
            text.append(text_clean)
    date = pd.to_datetime(date)
    df = pd.DataFrame([date, link_list, text]).T
    df.columns = ['date', 'link', 'text']
    return df

clean = clean_bdb(first_transcripts, bdblinks_519)

def take_monologue(transcript):
    if len(re.findall('sio:([^:]+)|yor:([^:]+)', str(transcript))) > 0:
        return str(re.findall('sio:([^:]+)|yor:([^:]+)', str(transcript)))
    else:
        return ''

toremove = "'\\,\"\[\(\]\)-–"

punc_lower = lambda x: re.sub('[%s]' % re.escape(toremove), '', x.lower())
remove_xa0 = lambda x: x.replace('xa0', '')
remove_space = lambda x: x.replace('  ', ' ')

clean['monologue'] = clean['text'].map(take_monologue).map(punc_lower).map(remove_xa0).map(remove_space)

# Include a column indicating this is a transcript from Mayor de Blasio
clean.insert(0, 'speaker', 'de blasio')

# with open('../data/bdbtranscript_519.pickle', 'wb') as to_write:
#     pickle.dump(clean, to_write)

## Clean Cuomo Transcripts

### Extract transcripts from source code

Get transcript date and text. The text requires multiple sequential cleaning steps, taking a more specific extract each time, and removing text that isn't relevant to the speech.

with open('../data/cuomosource_519.pickle', 'rb') as read_file:
    cuomosource_519 = pickle.load(read_file)
with open('../data/cuomolinks_519.pickle', 'rb') as read_file:
    cuomolinks_519 = pickle.load(read_file) 

def get_date(source_object):
    date_list = []
    for i in source_object:
        date = i.find('div', class_="published-date").text
        date_clean = re.search('\\n\\n(.+20.{2})', date).group(1).strip()
        date_list.append(date_clean)
    return date_list

cuomo_date = get_date(cuomosource_519)

def get_text(source_object):
    text_list = []
    for s in source_object:
        text = s.find('div', class_='field field--name-field-body field--type-text-long field--label-hidden')
        text_list.append(text)
    return text_list

cuomo_step1 = get_text(cuomosource_519)

def clean_cuomo(transcript_list):
    clean_transcripts = []
    for idx, i in enumerate(transcript_list):
        cleaned = remove_html_tags(str(i))
        if re.search('below:(.+)', cleaned) is not None:
            text_clean = re.search('below:(.+)', cleaned).group(1)
            clean_transcripts.append(text_clean)
        elif re.search('here.(.+)', cleaned) is not None:
            text_clean = re.search('here.(.+)', cleaned).group(1)
            clean_transcripts.append(text_clean)
    return clean_transcripts

cuomo_step2 = clean_cuomo(cuomo_step1)

I decided to check whether I was actually pulling transcripts for all of the pages, so I calculated the length for each transcript and looked at the shortest ones.

test_len = []
for i in cuomo_step2:
    test_len.append(len(i))
test_len_array = np.array(test_len)
test_len_array.argsort()

cuomo_step2[68]

This is not a valid transcript, so I will be removing this from the list. I'll look at the next shortest one, too.

cuomo_step2[99]

That one looks okay, so I will only delete the shortest transcript from my lists.

delete_index = test_len_array.argsort()[0]
del cuomo_step2[delete_index]
del cuomolinks_519[delete_index]
del cuomo_date[delete_index]

Now I will put them into a dataframe with the date, link, and transcript text, as I did for the de Blasio transcripts.

cuomo_date = pd.to_datetime(cuomo_date)

cuomo_clean = pd.DataFrame([cuomo_date, cuomolinks_519, cuomo_step2]).T
cuomo_clean.columns = ['date', 'links', 'text']

cuomo_clean

# with open('../data/cuomotranscript_519.pickle', 'wb') as to_write:
#     pickle.dump(cuomo_clean, to_write)

### Further Cleaning

I ended up needing to use three different functions using different regular expressions to extract just the parts of the transcript where Cuomo was speaking.

def take_monologue(transcript):
    if len(re.findall('Cuomo:([^:]+)', str(transcript))) > 0:
        return str(re.findall('Cuomo:([^:]+)', str(transcript)))
    else:
        return str(transcript)

def take_monologue2(transcript):
    if len(re.findall('Cuomo:(.*?)[A-Z][a-z]+:', str(transcript))) > 0:
        return str(re.findall('Cuomo:(.*?)[A-Z][a-z]+:', str(transcript)))
    else:
        return str(transcript)

def take_monologue3(transcript):
    if len(re.findall('^(.*?)[A-Z][a-z]+:', str(transcript))) > 0:
        return str(re.findall('^(.*?)[A-Z][a-z]+:', str(transcript)))
    else:
        return str(transcript)

cuomo_clean['monologue'] = cuomo_clean['text'].map(take_monologue)

cuomo_clean['monologue2'] = cuomo_clean['text'].map(take_monologue2)

# Only use third function if the first two don't capture the right transcript
def extract_m3(col1, col2):
    if (len(col1)-len(col2))/len(col2) > 10:
        return take_monologue3(col1)
    else:
        return ''

cuomo_clean['monologue3'] = cuomo_clean.apply(lambda x: extract_m3(x.text, x.monologue), axis = 1)

I want to keep the longer text, and supplement with the text taken from the third function if it exists.

def compare_two(col1, col2):
    if (len(col1)-len(col2))/len(col2) > 1:
        return col1
    else:
        return col2

cuomo_clean['final_text'] = cuomo_clean.apply(lambda x: compare_two(x.monologue, x.monologue2), axis = 1)

cuomo_clean['final_text2'] = cuomo_clean.monologue3 + cuomo_clean.final_text

remove_sxa0 = lambda x: x.replace('\xa0', '')

cuomo_clean['final_clean'] = cuomo_clean['final_text2'].map(punc_lower).map(remove_xa0).map(remove_sxa0).map(remove_space)

cuomo_final = cuomo_clean.loc[:, ['date', 'links', 'text', 'final_clean']]
cuomo_final.columns = ['date', 'link', 'text', 'monologue']

# Include a column indicating this is a transcript from Governor Cuomo
cuomo_final.insert(0, 'speaker', 'cuomo')

cuomo_final.head()

# with open('../data/cuomotranscript_519.pickle', 'wb') as to_write:
#     pickle.dump(cuomo_final, to_write)
