==> analysis/03_combine_speakers.ipynb <==
# Notebook 3: Combine Transcripts for Mayor de Blasio and Governor Cuomo

### Introduction

This is a very short notebook, focusing on combining the cleaned transcript data from both speakers created in Notebook 2.

import pickle
import pandas as pd
import numpy as np

with open('../data/cuomotranscript_519.pickle','rb') as read_file:
    cuomo = pickle.load(read_file)
with open('../data/bdbtranscript_519.pickle','rb') as read_file:
    deblasio = pickle.load(read_file)  

### Combine Transcripts

ny = pd.concat([cuomo, deblasio], axis = 0).reset_index(drop=True)

ny.info()

### Pickle Resulting Dataframe

# with open('../data/combined.pickle', 'wb') as to_write:
#     pickle.dump(ny, to_write)
